move database into multiple files, friends database
